- Actor Get, Create, Delete implemented

+++ src/main/java/javaproject/movies/domain/Genre.java
@@ -16,7 +16,7 @@ public class Genre {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long genreId;
+    private Integer genreId;
 
     private String name;
 }
