+++ src/main/java/javaproject/movies/domain/Review.java
@@ -18,7 +18,7 @@ public class Review {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long reviewId;
+    private Integer reviewId;
 
     @ManyToOne
     @JoinColumn(name = "movie_id")
